Combines the add and edit event forms and adds the form to the edit modal.
