+++ client/src/pages/pages.css
@@ -61,6 +61,7 @@ body::-webkit-scrollbar { /* WebKit */
 .attendee-list-shadow {
   padding-bottom: 2em;
 }
+
 @media only screen and (max-width: 768px) {
   .masonry-box{
     display: block;
